requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres==6.2.1
apache-airflow-providers-http==5.3.2
psycopg2-binary==2.9.10
pandas==2.2.3
requests==2.32.3
python-dotenv==1.1.1
pytest==8.4.1
pyarrow==20.0.0
